+++ web/src/js/ads/displayAd.js
@@ -1,6 +1,6 @@
 
 import adsEnabled from './adsEnabledStatus'
-import googleDisplayAd from './googleDisplayAd'
+import googleDisplayAd from './google/googleDisplayAd'
 import mockDisplayAd from './mockDisplayAd'
 
 export default function (adId) {
