+++ web/src/js/ads/google/__tests__/googleDisplayAd.test.js
