+++ web/src/js/ads/prebid/prebidConfig.js
@@ -1,5 +1,5 @@
 
-import { getGoogleTag } from '../googleTag'
+import { getGoogleTag } from '../google/googleTag'
 import { getPrebidPbjs } from './getPrebidPbjs'
 
 export default function () {
